+++ dialect/psql/table.go
@@ -18,11 +18,11 @@ type setter[T any] interface {
 	orm.Setter[T, *dialect.InsertQuery, *dialect.UpdateQuery]
 }
 
-func NewTable[T orm.Model, Tset setter[T]](schema, tableName string) *Table[T, []T, Tset] {
+func NewTable[T any, Tset setter[T]](schema, tableName string) *Table[T, []T, Tset] {
 	return NewTablex[T, []T, Tset](schema, tableName)
 }
 
-func NewTablex[T orm.Model, Tslice ~[]T, Tset setter[T]](schema, table string) *Table[T, Tslice, Tset] {
+func NewTablex[T any, Tslice ~[]T, Tset setter[T]](schema, table string) *Table[T, Tslice, Tset] {
 	setMapping := mappings.GetMappings(reflect.TypeOf((*new(Tset))))
 	view, mappings := newView[T, Tslice](schema, table)
 	t := &Table[T, Tslice, Tset]{
@@ -37,7 +37,7 @@ func NewTablex[T orm.Model, Tslice ~[]T, Tset setter[T]](schema, table string) *
 
 // The table contains extract information from the struct and contains
 // caches ???
-type Table[T orm.Model, Tslice ~[]T, Tset setter[T]] struct {
+type Table[T any, Tslice ~[]T, Tset setter[T]] struct {
 	*View[T, Tslice]
 	pkCols           orm.Columns
 	setterMapping    mappings.Mapping